requirements.txt
Twisted==14.0.0
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
pyserial==2.7
requests==2.4.3
tornado==4.0.2
zope.interface==4.1.1
RPi.GPIO==0.5.5
